+++ src/features/auth/routers/auth-router.ts
@@ -3,13 +3,13 @@ import { getSessionHandler } from '../route-handlers/get-session-handler';
 import { loginHandler } from '../route-handlers/login-handler';
 
 import { logoutHandler } from '../route-handlers/logout-handler';
-import { patchSessionHandler } from '../route-handlers/patch-session.handler';
+import { patchSessionHandler } from '../route-handlers/patch-session-handler';
 import { registerHandler } from '../route-handlers/register-handler';
 import { resetPasswordHandler } from '../route-handlers/reset-password-handler';
 import { sendPasswordResetEmailHandler } from '../route-handlers/send-password-reset-email-handler';
 
 import { sendVerificationEmailHandler } from '../route-handlers/send-verification-email';
-import { checkAuth } from '../util/check-auth';
+import { checkAuth } from '../middleware/check-auth';
 
 const router = getRouter();
 router.post('/login', loginHandler);